requirements.txt
torch==1.12.0
torchvision==0.13.0
Pillow==10.2.0
numpy==1.26.4
pandas==2.1.3
matplotlib==3.8.3
scipy==1.7.3
image-similarity-measures==0.3.6
scikit-learn==1.4.0
tqdm==4.66.1
PyYAML==6.0.1
